requirements.txt
burr[tracking,cli,graphviz]==0.33.0
ollama==0.3.3
graphviz==0.20.3
PyAudio==0.2.14
SpeechRecognition==3.11.0
nltk==3.9.1
numpy==2.1.3
pyttsx3==2.98
fuzzywuzzy==0.18.0
pyalsaaudio==0.11.0
vosk==0.3.45
soundfile==0.12.1
openai==1.57.2
pvporcupine==3.0.3
websocket==0.2.1
python-dotenv==1.0.1
tqdm==4.67.0
pydub==0.25.1
websocket_client==1.8.0
bytewax==0.21.1